+++ independent-projects/resteasy-reactive/common/runtime/src/main/java/org/jboss/resteasy/reactive/common/core/BlockingOperationSupport.java
@@ -1,9 +1,5 @@
-package org.jboss.resteasy.reactive.server.core;
+package org.jboss.resteasy.reactive.common.core;
 
-/**
- * @deprecated This class has been replaced by {@link org.jboss.resteasy.reactive.common.core.BlockingOperationSupport}
- */
-@Deprecated(forRemoval = true, since = "3.27")
 public class BlockingOperationSupport {
 
     private static volatile IOThreadDetector ioThreadDetector;
